requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
airflow-provider-duckdb==0.2.0
astronomer-cosmos==1.0.4
airflow-dbt==0.4.0
pandas==2.2.2
duckdb==0.9.2